MTL-1828 Remove manual chrony steps and cleanup deploy steps (#2977)
* MTL-1828 Remove manual step for NTP pool

The NTP default pool is now removed from the image in the pipeline,
removing the need for a user to do this themselves.

* Consolidate and remove non-deploy steps

After MTL-1828's commit, the Cleanup step contains a step for installing
the validation tests and that step makes more sense underneath the
"Validate deployment" section. The relocation of that step empties the
"Cleanup" step, and it can be deleted.

This also removes the LVM checks, removing the scripts from the docs
repo (reducing our code footprint in a documentation repo). The LVM
checks are done by csm-testing, and the UX shouldn't have the user run
these scripts during validation. The better UX here is 1-2-3; setup,
deploy, validate (vs. setup, deploy, run ad-hoc, validate).

This also gives a linkable header to the `kubectl` config setup step,
making it easier to jump to the step and for calling it out.