CASMINST-5767: fix variable construction, add missing vars (#3006)
* CASMINST-5767: fix variable construction, add missing vars

The KUBERNETES_VERSION and CEPH_VERSION variables were mistakenly
removed in a previous PR. They are still needed to ensure we
collect the correct squashfs files out of the CSM tarball.

The ${arch} variable was missing in a conditional, causing the
script to always invoke ncn-image-modification.sh even if it had
already run successfully on a previous invocation.

If a non-relative path containing a file glob (*) is quoted, bash
will add single quotes to the whole expression. This caused the
path - with the glob - to be evaluated literally (without expanding
the glob). This resulted in a path being "not found".

* Update upgrade/scripts/upgrade/prerequisites.sh

Co-authored-by: Russell Bunch <doomslayer@hpe.com>
Signed-off-by: John Heemstra <heemstra@users.noreply.github.com>

* Update upgrade/scripts/upgrade/prerequisites.sh

Co-authored-by: Russell Bunch <doomslayer@hpe.com>
Signed-off-by: John Heemstra <heemstra@users.noreply.github.com>

Signed-off-by: John Heemstra <heemstra@users.noreply.github.com>
Co-authored-by: Russell Bunch <doomslayer@hpe.com>