CASMINST-5812: Split prepare-images and update-cfs-config IUF stages (#3039)
Split the `prepare-images` and `update-cfs-config` IUF stages into two
operations, one for management nodes and one for managed nodes. The IUF
CLI accepts separate options for specifying the bootprep file for the
management nodes vs. the managed nodes. The `sat bootprep` command
expects to receive only on input file per invocation. As such, we would
already have to call it twice. Splitting the two invocations into
separate operations will make it easier for the subsequent stages,
`management_nodes_rollout` and `managed_nodes_rollout`, to obtain the
output describing the CFS configurations, IMS images, and BOS session
templates that were created for the type of nodes being rolled out.

Note that currently, the IUF runs operations serially within a stage.
There is a Jira (CASMINST-5702) to improve this so that operations in a
stage can be run in parallel. This will significantly speed up the
`prepare-images` stage by creating management node and managed node
images in parallel.